requirements.txt
langchain
langchain_community
pypdf
sentence-transformers
faiss-cpu
streamlit 
google-generativeai
pyngrok